// packages/studio-ui/src/types.ts
export interface DefaultNodeProps {
  [prop: string]: unknown;
}
// ...
export interface StudioStateDefinition {
  name: string;
  initialValue: any;
  // TODO: state type?
}

export interface StudioBoundProp {
  type: 'binding';
  state: string;
}

export interface StudioConstantProp<V> {
  type: 'const';
  value: V;
}

export type StudioNodeProp<V> = StudioConstantProp<V> | StudioBoundProp;

export type StudioNodeProps<P = DefaultNodeProps> = {
  readonly [K in keyof P]: StudioNodeProp<P[K]> | undefined;
};

export type NodeId = Branded<string, 'NodeId'>;

export interface StudioNode<P = DefaultNodeProps> {
  readonly id: NodeId;
  readonly parentId: NodeId | null;
  readonly component: string;
  readonly name: string;
  readonly props: Partial<StudioNodeProps<P>>;
  readonly children: NodeId[];
}

export interface StudioPageSummary {
  id: string;
}

export interface GridSlot {
  span: number;
  content: NodeId | null;
}

export interface StudioNodes {
  [id: NodeId]: StudioNode | undefined;
}

export interface StudioPageQuery<Q> {
  connectionId: string;
  query: Q;
}

export interface StudioPage extends StudioPageSummary {
  nodes: StudioNodes;
  root: NodeId;
  state: Record<string, StudioStateDefinition>;
  queries: {
    [id: string]: StudioPageQuery<any> | undefined;
  };
}